Updated jumbotron and welcome callout separation, fixed bottom margin of callout
