+++ cordapp/frontend/src/pages/HomePage.js
@@ -2,6 +2,7 @@ import React from "react";
 import { useAuth } from "../auth-hook";
 import { Widgets } from "../pages/views/home/Widgets";
 import { Jumbotron } from "../pages/views/home/Jumbotron";
+import { WelcomeCallout } from "../pages/views/home/WelcomeCallout";
 
 const HomePage = () => {
   const auth = useAuth();
@@ -9,6 +10,7 @@ const HomePage = () => {
   return (
     <>
       <Jumbotron auth={auth} />
+      <WelcomeCallout auth={auth} />
       <Widgets auth={auth} />
     </>
   );
