+++ cordapp/frontend/src/pages/views/home/Jumbotron.js
@@ -1,37 +1,7 @@
-import React, { useState, useEffect } from "react";
-import { CCol, CRow, CJumbotron, CButton, CCallout } from "@coreui/react";
+import React from "react";
+import { CCol, CRow, CJumbotron, CButton } from "@coreui/react";
 
 export const Jumbotron = ({ auth }) => {
-  const [hour, setHour] = useState();
-  const [greeting, setGreeting] = useState();
-
-  useEffect(() => {
-    const fetchGreeting = async () => {
-      const date = new Date();
-      setHour(date.getHours());
-
-      if (hour) {
-        switch (true) {
-          case hour < 12:
-            setGreeting("Good morning, ");
-            break;
-          case hour > 12 && hour < 17:
-            setGreeting("Good afternoon, ");
-            break;
-          case hour >= 17:
-            setGreeting("Good evening, ");
-            break;
-          default:
-            setGreeting("Hello, ");
-            break;
-        }
-      }
-    };
-    if (auth.isAuthenticated) {
-      fetchGreeting();
-    }
-  }, [hour, greeting, auth]);
-
   return (
     <CRow>
       <CCol>
@@ -51,15 +21,6 @@ export const Jumbotron = ({ auth }) => {
               </CCol>
             </CRow>
           </CJumbotron>
-        )}
-        {auth.isAuthenticated && (
-          <CCallout color="dark" className="bg-light mt-0">
-            <h1 className="text-muted">
-              {greeting}
-              {auth.user.firstName}!
-            </h1>
-            <h6>Your daily summary:</h6>
-          </CCallout>
         )}
       </CCol>
     </CRow>
